WebSite/requirements.txt
arrow==0.12.1
certifi==2018.11.29
chardet==3.0.4
confusable-homoglyphs==3.2.0
Django==2.1.4
django-debug-toolbar==1.11
django-extra-views==0.12.0
django-registration==3.0
django-requestlogging-redux==1.2.1
django-widget-tweaks==1.4.3
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
pip==18.1
psycopg2==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
reportlab==3.5.12
requests==2.21.0
setuptools==40.6.3
six==1.12.0
sqlparse==0.2.4
stripe==2.17.0
urllib3==1.24.1
